[feat]新增UUID处理逻辑

+++ amt630hV100-sdk-beta-20250625/amt630hv100-freertos/app/hcn/common/utils/hcn_utils.h
@@ -38,7 +38,7 @@ typedef struct {
  * @return 无
  */
 void hcn_hex_config_data_print(char const *function, char *prefix, uint8_t *data,
-                              uint8_t length);
+                              uint16_t length);
 
 /**
  * @brief  字符串转小写
